chores: extra module for config_importer plus refactoring
